% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioInstaller.R
\docType{package}
\name{BioInstaller}
\alias{BioInstaller}
\alias{BioInstaller-package}
\title{Install and download massive bioinformatics analysis software and database,
such as NGS analysis tools with its required database or/and reference genome,
is still a complicated task. BioInstaller can be used to install
that tools and database in R conveniently.}
\description{
Install and download massive bioinformatics analysis software and database,
such as NGS analysis tools with its required database or/and reference genome,
is still a complicated task. BioInstaller can be used to install
that tools and database in R conveniently.
}
\seealso{
Useful links:

\url{https://github.com/Miachol/BioInstaller}

Report bugs at \url{https://github.com/Miachol/BioInstaller/issues}
}
\author{
Li Jianfeng \url{lee_jianfeng@sjtu.edu.cn}
}
