% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{change.info}
\alias{change.info}
\title{Update biologly softwares infomation of system}
\usage{
change.info(name = "", installed = TRUE, source.dir = "", bin.dir = "",
  executable.files = "", db = Sys.getenv("BIO_SOFTWARES_DB_ACTIVE",
  system.file("extdata", "softwares_db_demo.yaml", package = "BioInstaller")),
  ..., verbose = FALSE)
}
\arguments{
\item{name}{Software name}

\item{installed}{Wheather be installed successful in system}

\item{source.dir}{Directorie of softwares source code}

\item{bin.dir}{Directorie of softwares bin}

\item{executable.files}{Executable files in bin.dir}

\item{db}{File of saving softwares infomation}

\item{...}{Other key and value paired need be saved in BioInstaller}

\item{verbose}{TRUE is debug mode}
}
\value{
Bool Value
}
\description{
Update biologly softwares infomation of system
}
\examples{
set.biosoftwares.db(sprintf('\%s/.BioInstaller', tempdir()))
change.info(name = 'demo', installed = 'yes', source.dir = '',
bin.dir = '', excutable.files = c('demo'), others.customer = 'demo')
}
