% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{check_shiny_dep}
\alias{check_shiny_dep}
\title{Function to check dependence package running BioInstaller shiny app}
\usage{
check_shiny_dep(install = FALSE)
}
\arguments{
\item{install}{Logical wheather to install the shiny app depence package}
}
\description{
Function to check dependence package running BioInstaller shiny app
}
\examples{
\dontrun{
check_shiny_dep()
}
}
