% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda.env.create}
\alias{conda.env.create}
\title{Wrapper function of 'conda env create', create an environment based on an environment file}
\usage{
conda.env.create(env_name = "", env_file = "", env_path = "",
  params = "", ...)
}
\arguments{
\item{env_name}{Name of environment}

\item{env_file}{Environment definition file (default: environment.yml)}

\item{env_path}{Full path to environment prefix}

\item{params}{Extra command line parameters of conda}

\item{...}{Parameters pass to \code{\link{conda}}}
}
\description{
Wrapper function of 'conda env create', create an environment based on an environment file
}
\examples{
\dontrun{
  conda.env.create(params = 'vader/deathstar')
  conda.env.create(env_name = 'name')
  conda.env.create(env_file = '/path/to/environment.yml')
  conda.env.create(env_name = 'deathstar', 
                   env_file = '/path/to/requirements.txt')
  conda.env.create(env_file = '/path/to/requirements.txt', 
  env_path = '/home/user/software/deathstar')
}
}
