% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda.env.list}
\alias{conda.env.list}
\title{Wrapper function of 'conda env list', list the Conda environments}
\usage{
conda.env.list(...)
}
\arguments{
\item{...}{Parameters pass to \code{\link{conda}}}
}
\description{
Wrapper function of 'conda env list', list the Conda environments
}
\examples{
\dontrun{
  conda.env.list()
}
}
