% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda.list}
\alias{conda.list}
\title{Wrapper function of 'conda list', list linked packages in a conda environment.}
\usage{
conda.list(env_name = "base", ...)
}
\arguments{
\item{env_name}{Name of environment, default is current}

\item{...}{Parameters pass to \code{\link{conda}}}
}
\description{
Wrapper function of 'conda list', list linked packages in a conda environment.
}
\examples{
\dontrun{
  conda.list()
  conda.list(env_name = 'your_env')
}
}
