% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda}
\alias{conda}
\title{Wrapper function of conda}
\usage{
conda(suffix_params = "", prefix_params = "",
  conda = Sys.which("conda"), ...)
}
\arguments{
\item{suffix_params}{Command line parameters of conda}

\item{prefix_params}{Command line parameters of conda}

\item{conda}{Default is Sys.which('conda')}

\item{...}{Parameters pass to 'system'}
}
\description{
Wrapper function of conda
}
\examples{
\dontrun{
  conda()
}
}
