% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{copy_plugins}
\alias{copy_plugins}
\title{Function to copy the default plugins of BioInstaller}
\usage{
copy_plugins(plugin_dir = "~/.BioInstaller/plugins",
  template_dir = system.file("extdata", "config/shiny/", package =
  "BioInstaller"), pattern = "shiny.*.parameters.toml", auto_create = FALSE)
}
\arguments{
\item{plugin_dir}{The destdir to store plugins [~/.BioInstaller/plugins]}

\item{template_dir}{The template dir system.file('extdata', 'config/shiny/', 
package = 'BioInstaller')}

\item{pattern}{Used in \code{\link{list.files}} ["shiny.*.parameters.toml"]}

\item{auto_create}{Auto create dir, default is FALSE}
}
\description{
Function to copy the default plugins of BioInstaller
}
\examples{
copy_plugins(tempdir())
}
