% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{craw.all.versions}
\alias{craw.all.versions}
\title{A function can be used to craw all source code from nongithub.cfg stored information}
\usage{
craw.all.versions(name, download.dir = "./",
  nongithub.cfg = system.file("extdata", "nongithub.toml", package =
  "BioInstaller"), ...)
}
\arguments{
\item{name}{Software name}

\item{download.dir}{Source code stored dir}

\item{nongithub.cfg}{Configuration file of installed by non github url,}

\item{...}{Other parameters pass to \code{\link[configr]{parse.extra}}
default is system.file('extdata', 'nongithub.toml', package='BioInstaller')}
}
\description{
A function can be used to craw all source code from nongithub.cfg stored information
}
\examples{
craw.all.versions('demo')
}
