% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{crawl.all.versions}
\alias{crawl.all.versions}
\title{A function can be used to craw all source code from nongithub.cfg stored information}
\usage{
crawl.all.versions(name, download.dir = "./",
  nongithub.cfg = c(system.file("extdata",
  "config/nongithub/nongithub.toml", package = "BioInstaller"),
  system.file("extdata", "config/db/db_main.toml", package =
  "BioInstaller"), system.file("extdata", "config/db/db_annovar.toml",
  package = "BioInstaller"), system.file("extdata",
  "config/db/db_blast.toml", package = "BioInstaller")),
  parse.extra.params = list(extra.list = list(), rcmd.parse = TRUE,
  bash.parse = TRUE, glue.parse = TRUE), license = "")
}
\arguments{
\item{name}{Software name}

\item{download.dir}{Download destdir}

\item{nongithub.cfg}{Configuration file of installed by non github url,
default is system.file('extdata', 'config/nongithub/nongithub.toml', package='BioInstaller')}

\item{parse.extra.params}{Other parameters pass to \code{\link[configr]{parse.extra}}}

\item{license}{The BioInstaller download license code.}
}
\description{
A function can be used to craw all source code from nongithub.cfg stored information
}
\examples{
crawl.all.versions('demo')
}
