% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker.search}
\alias{docker.search}
\title{Search softwares docker infomation in BioInstaller docker database}
\usage{
docker.search(name, docker.db = system.file("extdata",
  "config/docker/docker.toml", package = "BioInstaller"))
}
\arguments{
\item{name}{Software name, e.g bwa}

\item{docker.db}{A list including docker repo infomation, 
default to use built-in config/docker/docker.toml}
}
\value{
A list
}
\description{
Search softwares docker infomation in BioInstaller docker database
}
\examples{
docker.search('bwa')
}
