% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{is.biosoftwares.db.active}
\alias{is.biosoftwares.db.active}
\title{Test active configuration file}
\usage{
is.biosoftwares.db.active(biosoftwares.db)
}
\arguments{
\item{biosoftwares.db}{Configuration filename of bio-softwares db}
}
\value{
Logical indicating whether the specified configuration file is active
}
\description{
Check whether a Bio Softwares DB is active
}
\examples{
is.biosoftwares.db.active('config.cfg')
}
