% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{new.bioinfo}
\alias{new.bioinfo}
\title{Create new BioInstaller items to github forum}
\usage{
new.bioinfo(config.file = "github.toml", title = "", description = "",
  publication = "")
}
\arguments{
\item{config.file}{github.toml, nongithub.toml, db_annovar.toml, db_main.toml, or new}

\item{title}{Name of new item}

\item{description}{Description of new item}

\item{publication}{Publication of new item}
}
\description{
Create new BioInstaller items to github forum
}
\examples{
new.bioinfo('db_main.toml', 'test_item', 'Just is a test item', 'NA')
}
