% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{set.biosoftwares.db}
\alias{set.biosoftwares.db}
\title{Set BIO_SOFWARES_DB_ACTIVE as the BioInstaller db}
\usage{
set.biosoftwares.db(biosoftwares.db)
}
\arguments{
\item{biosoftwares.db}{Configuration filename of bio-softwares db}
}
\value{
Logical indicate wheather set db successful
}
\description{
Set BIO_SOFWARES_DB_ACTIVE as the BioInstaller db
}
\examples{
set.biosoftwares.db(sprintf('\%s/.BioInstaller', tempdir()))
}
