% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{show.installed}
\alias{show.installed}
\title{Show all installed bio-softwares in system}
\usage{
show.installed(db = Sys.getenv("BIO_SOFTWARES_DB_ACTIVE",
  system.file("extdata", "softwares_db_demo.yaml", package = "BioInstaller")),
  only.installed = TRUE, verbose = FALSE)
}
\arguments{
\item{db}{File of saving softwares infomation}

\item{only.installed}{Logical wheather only show installed softwares in db}

\item{verbose}{TRUE is debug mode}
}
\value{
Bool Value
}
\description{
Show all installed bio-softwares in system
}
\examples{
set.biosoftwares.db(sprintf('\%s/.BioInstaller', tempdir()))
show.installed()
}
