% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spack.R
\name{spack.list}
\alias{spack.list}
\title{Wrapper function of 'spack list', list and search available packages}
\usage{
spack.list(...)
}
\arguments{
\item{...}{Parameters pass to \code{\link{spack}}}
}
\description{
Wrapper function of 'spack list', list and search available packages
}
\examples{
\dontrun{
  spack.list()
}
}
