% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spack.R
\name{spack}
\alias{spack}
\title{Wrapper function of spack}
\usage{
spack(suffix_params = "", prefix_params = "",
  spack = Sys.which("spack"), ...)
}
\arguments{
\item{suffix_params}{Command line parameters of spack (prefix_params spack suffix_params)}

\item{prefix_params}{Command line parameters of spack (prefix_params spack suffix_params)}

\item{spack}{Default is Sys.which('spack')}

\item{...}{Parameters pass to 'system'}
}
\description{
Wrapper function of spack
}
\examples{
\dontrun{
  spack()
}
}
