% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioInstaller.R
\docType{package}
\name{BioInstaller}
\alias{BioInstaller}
\alias{BioInstaller-package}
\title{This package is a new platform to construct interactive and reproducible biological 
data analysis applications based on R language, which includes the R functions and R 
Shiny application, REST APIs.}
\description{
This package is a new platform to construct interactive and reproducible biological 
data analysis applications based on R language, which includes the R functions and R 
Shiny application, REST APIs.
}
\seealso{
Useful links:

\url{https://github.com/JhuangLab/BioInstaller}

Report bugs at \url{https://github.com/JhuangLab/BioInstaller/issues}
}
\author{
Li Jianfeng \url{lee_jianfeng@sjtu.edu.cn}
}
