% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioInstaller.R
\docType{package}
\name{BioInstaller}
\alias{BioInstaller}
\alias{BioInstaller-package}
\title{This package is a comprehensive R package to integrate 
bioinformatics resources, such as software/script and database}
\description{
This package is a comprehensive R package to integrate 
bioinformatics resources, such as software/script and database
}
\seealso{
Useful links:

\url{https://github.com/JhuangLab/BioInstaller}

Report bugs at \url{https://github.com/JhuangLab/BioInstaller/issues}
}
\author{
Li Jianfeng \url{lee_jianfeng@sjtu.edu.cn}
}
