% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioInstaller.R
\docType{package}
\name{BioInstaller}
\alias{BioInstaller}
\alias{BioInstaller-package}
\title{This package can be used to install and download massive bioinformatics analysis 
softwares and databases, such as NGS reads mapping tools with its required databases.}
\description{
This package can be used to install and download massive bioinformatics analysis 
softwares and databases, such as NGS reads mapping tools with its required databases.
}
\seealso{
Useful links:

\url{https://github.com/JhuangLab/BioInstaller}

Report bugs at \url{https://github.com/JhuangLab/BioInstaller/issues}
}
\author{
Li Jianfeng \url{lee_jianfeng@sjtu.edu.cn}
}
