% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{copy_configs}
\alias{copy_configs}
\title{Function to copy the default configuration file of BioInstaller}
\usage{
copy_configs(config_dir = "~/.BioInstaller/",
  template_dir = Sys.getenv("BIOINSTALLER_SHINY_CONFIG",
  system.file("extdata", "config/shiny/", package = "BioInstaller")),
  pattern = "shiny.config.yaml", auto_create = FALSE)
}
\arguments{
\item{config_dir}{The destdir to store plugins [~/.BioInstaller]}

\item{template_dir}{The template dir system.file('extdata', 'config/shiny/', 
package = 'BioInstaller')}

\item{pattern}{Used in \code{\link{list.files}} ['shiny.config.yaml']}

\item{auto_create}{Auto create dir, default is FALSE}
}
\description{
Function to copy the default configuration file of BioInstaller
}
\examples{
copy_configs(tempdir())
}
