% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{craw.all.versions}
\alias{craw.all.versions}
\title{A function can be used to craw all source code from nongithub.cfg stored information}
\usage{
craw.all.versions(name, download.dir = "./",
  nongithub.cfg = c(system.file("extdata", "config/nongithub/nongithub.toml",
  package = "BioInstaller"), system.file("extdata", "config/db/db_main.toml",
  package = "BioInstaller"), system.file("extdata", "config/db/db_annovar.toml",
  package = "BioInstaller"), system.file("extdata", "config/db/db_blast.toml",
  package = "BioInstaller")), parse.extra.params = list(rcmd.parse = TRUE,
  bash.parse = TRUE, glue.parse = TRUE))
}
\arguments{
\item{name}{Software name}

\item{download.dir}{Download destdir}

\item{nongithub.cfg}{Configuration file of installed by non github url,}

\item{parse.extra.params}{Other parameters pass to \code{\link[configr]{parse.extra}}
default is system.file('extdata', 'config/nongithub/nongithub.toml', package='BioInstaller')}
}
\description{
A function can be used to craw all source code from nongithub.cfg stored information
}
\examples{
craw.all.versions('demo')
}
