% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{del.info}
\alias{del.info}
\title{Delete biologly softwares infomation of system}
\usage{
del.info(name = "", db = Sys.getenv("BIO_SOFTWARES_DB_ACTIVE",
  system.file("extdata", "softwares_db_demo.yaml", package = "BioInstaller")),
  verbose = FALSE)
}
\arguments{
\item{name}{Software name}

\item{db}{File of saving softwares infomation}

\item{verbose}{TRUE is debug mode}
}
\value{
Bool Value
}
\description{
Delete biologly softwares infomation of system
}
\examples{
set.biosoftwares.db(sprintf('\%s/.BioInstaller', tempdir()))
del.info('bwa')
}
