% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get.info}
\alias{get.info}
\title{Show biologly softwares infomation of system}
\usage{
get.info(name = "", db = Sys.getenv("BIO_SOFTWARES_DB_ACTIVE",
  system.file("extdata", "demo/softwares_db_demo.yaml", package =
  "BioInstaller")), verbose = TRUE)
}
\arguments{
\item{name}{Software name}

\item{db}{File of saving softwares infomation}

\item{verbose}{Ligical indicating wheather show the log message}
}
\value{
Bool Value
}
\description{
Show biologly softwares infomation of system
}
\examples{
db <- sprintf('\%s/.BioInstaller', tempdir())
set.biosoftwares.db(db)
change.info(name = 'bwa', installed = 'yes', source.dir = '',
bin.dir = '', excutable.files = c('demo'), others.customer = 'demo')
get.info('bwa')
unlink(db)
}
