% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{show.installed}
\alias{show.installed}
\title{Show all installed bio-softwares in system}
\usage{
show.installed(db = Sys.getenv("BIO_SOFTWARES_DB_ACTIVE", tempfile()),
  only.installed = TRUE, verbose = TRUE)
}
\arguments{
\item{db}{File saving softwares infomation}

\item{only.installed}{Logical wheather only show installed softwares in db}

\item{verbose}{Ligical indicating wheather show the log message}
}
\value{
Bool Value
}
\description{
Show all installed bio-softwares in system
}
\examples{
db <- sprintf('\%s/.BioInstaller', tempdir())
set.biosoftwares.db(db)
change.info(name = 'bwa', installed = 'yes', source.dir = '',
bin.dir = '', excutable.files = c('demo'), others.customer = 'demo')
show.installed()
unlink(db)
}
