% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{web}
\alias{web}
\title{Function to run BioInstaller shiny APP service}
\usage{
web(appDir = system.file("extdata", "shiny", package = "BioInstaller"),
  auto_create = FALSE, ...)
}
\arguments{
\item{appDir}{The application to run.
Default is system.file('extdata', 'tools/shiny/R', package = 'BioInstaller')}

\item{auto_create}{Auto create dir, default is FALSE}

\item{...}{Other parameters pass to \code{\link[shiny]{runApp}}}
}
\description{
Function to run BioInstaller shiny APP service
}
\examples{
\dontrun{
  web(auto_create = TRUE)
}
}
